refactor: move library/util code to lib/resources/mail
