+++ src/commands/mail/deliverybox/create.tsx
@@ -11,7 +11,7 @@ import { Success } from "../../../rendering/react/components/Success.js";
 import { ReactNode } from "react";
 import { ProcessRenderer } from "../../../rendering/process/process.js";
 import { FlagInput, OutputFlags } from "@oclif/core/lib/interfaces/parser.js";
-import { generateRandomPassword } from "../commons.js";
+import { generateRandomPassword } from "../../../lib/resources/mail/commons.js";
 
 type CreateResult = {
   deliveryBoxId: string;
