+++ src/commands/mail/deliverybox/update.tsx
@@ -13,7 +13,7 @@ import {
   mailDeliveryBoxArgs,
   withDeliveryBoxId,
 } from "../../../lib/resources/mail/flags.js";
-import { generateRandomPassword } from "../commons.js";
+import { generateRandomPassword } from "../../../lib/resources/mail/commons.js";
 
 type UpdateResult = {
   generatedPassword: string | null;
